Renamed project and files to birthday card
