+++ app/src/androidTest/java/com/example/birthdayCard/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.starter
+package com.example.birthdayCard
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.starter", appContext.packageName)
+        assertEquals("com.example.birthdaycard", appContext.packageName)
     }
 }
