+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:tools="http://schemas.android.com/tools"
+    package="com.example.birthdaycard">
 
     <application
         android:allowBackup="true"
@@ -10,13 +11,13 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.Starter"
+        android:theme="@style/Theme.BirthdayCard"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".BirthdayCardActivity"
             android:exported="true"
             android:label="@string/app_name"
-            android:theme="@style/Theme.Starter">
+            android:theme="@style/Theme.BirthdayCard">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
