+++ app/src/main/java/com/example/birthdaycard/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.example.starter.ui.theme
+package com.example.birthdayCard.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
